// app/backend/src/interfaces/drinks/IDrink.ts
export interface IDrink {
  idDrink: number;
  strDrink: string;
  strDrinkAlternate: string;
  strTags: string;
  strVideo: string;
  idCategory: number;
  strIba: string;
  strAlcoholic: string;
  strGlass: string;
  strInstructions: string;
  strDrinkThumb: string;
  strImageSource: string;
  strImageAttribution: string;
  strCreativeCommonsConfirmed: string;
  dateModified: Date | null;
}